+++ chalice/deploy/deployer.py
@@ -35,7 +35,6 @@ from chalice.constants import DEFAULT_STAGE_NAME, LAMBDA_TRUST_POLICY
 from chalice.constants import DEFAULT_LAMBDA_TIMEOUT
 from chalice.constants import DEFAULT_LAMBDA_MEMORY_SIZE
 from chalice.constants import MAX_LAMBDA_DEPLOYMENT_SIZE
-from chalice.constants import DEFAULT_APIGATEWAY_STAGE_NAME
 from chalice.policy import AppPolicyGenerator
 
 
@@ -805,8 +804,7 @@ class APIGatewayDeployer(object):
         # for each rest API, but that would require injecting chalice stage
         # information into the swagger generator.
         rest_api_id = self._aws_client.import_rest_api(swagger_doc)
-        api_gateway_stage = (config.api_gateway_stage or
-                             DEFAULT_APIGATEWAY_STAGE_NAME)
+        api_gateway_stage = config.api_gateway_stage
         self._deploy_api_to_stage(rest_api_id, api_gateway_stage,
                                   deployed_resources)
         return rest_api_id, self._aws_client.region_name, api_gateway_stage
@@ -819,8 +817,7 @@ class APIGatewayDeployer(object):
         LOGGER.debug("Generating swagger document for rest API.")
         swagger_doc = generator.generate_swagger(config.chalice_app)
         self._aws_client.update_api_from_swagger(rest_api_id, swagger_doc)
-        api_gateway_stage = (config.api_gateway_stage or
-                             DEFAULT_APIGATEWAY_STAGE_NAME)
+        api_gateway_stage = config.api_gateway_stage
         self._deploy_api_to_stage(
             rest_api_id, api_gateway_stage,
             deployed_resources)