+++ tests/unit/deploy/test_deployer.py
@@ -98,6 +98,7 @@ def config_obj(sample_app):
     config = Config.create(
         chalice_app=sample_app,
         stage='dev',
+        api_gateway_stage='api',
     )
     return config
 
@@ -140,7 +141,7 @@ def test_api_gateway_deployer_redeploy_api(config_obj, ui):
     # The rest_api_id does not exist which will trigger
     # the initial import
     deployed = DeployedResources(
-        None, None, None, 'existing-id', 'dev', None, None, {})
+        None, None, None, 'existing-id', 'api', None, None, {})
     aws_client.rest_api_exists.return_value = True
     lambda_arn = 'arn:aws:lambda:us-west-2:account-id:function:func-name'
 
@@ -164,7 +165,7 @@ def test_api_gateway_deployer_delete(config_obj, ui):
 
     rest_api_id = 'abcdef1234'
     deployed = DeployedResources(
-        None, None, None, rest_api_id, 'dev', None, None, {})
+        None, None, None, rest_api_id, 'api', None, None, {})
     aws_client.rest_api_exists.return_value = True
 
     d = APIGatewayDeployer(aws_client, ui)
@@ -178,7 +179,7 @@ def test_api_gateway_deployer_delete_already_deleted(ui):
     aws_client.delete_rest_api.side_effect = ResourceDoesNotExistError(
         rest_api_id)
     deployed = DeployedResources(
-        None, None, None, rest_api_id, 'dev', None, None, {})
+        None, None, None, rest_api_id, 'api', None, None, {})
     aws_client.rest_api_exists.return_value = True
     d = APIGatewayDeployer(aws_client, ui)
     d.delete(deployed)